REFACTOR: moved all base services to to `./common`  namespace
