+++ src/services/ContractService.ts
@@ -3,7 +3,7 @@ import { UserModel } from "@/models/UserModel";
 import { Inject, Service } from "@tsed/di";
 import { NotFound } from "@tsed/exceptions";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseCRUDService } from "./BaseCRUDService`1";
+import { BaseCRUDService } from "./common/BaseCRUDService`1";
 
 @Service()
 export class ContractService extends BaseCRUDService<ContractModel> {
