+++ src/services/CostBorrowerService.ts
@@ -2,7 +2,7 @@ import { AccountModel } from "@/models/AccountModel";
 import { CostBorrowerModel } from "@/models/CostBorrowerModel";
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseCostService } from "./BaseCostService";
+import { BaseCostService } from "./common/BaseCostService";
 
 @Service()
 export class CostBorrowerService extends BaseCostService<CostBorrowerModel> {
