+++ src/services/CreditTransactionService.ts
@@ -2,7 +2,7 @@ import { ContractModel } from "@/models/ContractModel";
 import { CreditContractTransactionModel } from "@/models/CreditContractTransactionModel";
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseContractTransactionService } from "./BaseContractTransactionService`1";
+import { BaseContractTransactionService } from "./common/BaseContractTransactionService`1";
 
 @Service()
 export class CreditTransactionService extends BaseContractTransactionService<CreditContractTransactionModel> {
