+++ src/services/DebitTransactionService.ts
@@ -2,7 +2,7 @@ import { ContractModel } from "@/models/ContractModel";
 import { DebitContractTransactionModel } from "@/models/DebitContractTransactionModel";
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseContractTransactionService } from "./BaseContractTransactionService`1";
+import { BaseContractTransactionService } from "./common/BaseContractTransactionService`1";
 
 @Service()
 export class DebitTransactionService extends BaseContractTransactionService<DebitContractTransactionModel> {
