+++ src/services/FixAmountPenaltyService.ts
@@ -1,7 +1,7 @@
 import { FixAmountPenaltyModel } from "@/models/FixAmountPenaltyModel";
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
-import { BaseCRUDService } from "./BaseCRUDService`1";
+import { BaseCRUDService } from "./common/BaseCRUDService`1";
 
 @Service()
 export class FixAmountPenaltyService extends BaseCRUDService<FixAmountPenaltyModel> {
