+++ src/services/RateOfBaseService.ts
@@ -1,7 +1,7 @@
 import { Inject, Service } from "@tsed/di";
 import { MongooseModel } from "@tsed/mongoose";
 import { RateOfBasePenaltyModel } from "../models/RateOfBasePenaltyModel";
-import { BaseCRUDService } from "./BaseCRUDService`1";
+import { BaseCRUDService } from "./common/BaseCRUDService`1";
 
 @Service()
 export class RateOfBaseService extends BaseCRUDService<RateOfBasePenaltyModel> {
